Initial commit
Hello! This is a reverse Polish notation REPL written in TypeScript with
Node as the target. TypeScript was chosen to provide type safety* on top
of JavaScript, which has none, in addition to other nice features. Style
is enforced with ESLint, helping to keep our code consistent. We use the
Airbnb style here, which is one of the most popular, in addition to some
tweaks that are listed and justified below. The program is just a prompt
that echos its input for this initial commit; the parsing and evaluation
come later!

brace-style:
    This project adopts the Stroustrup block style. In all popular style
    configurations, the else block is singled out as the only block that
    is different from all others. This impacts style consistency and the
    code's readability by offsetting else blocks from the left-hand side
    where one visually scans. We fix these issues with this choice.

no-console:
    This rule is turned off. You need the console for a CLI program. :-)

no-else-return:
    This rule is turned off. The style of demanding the omission of else
    and requiring an if statement fall through makes spotting cyclomatic
    complexity harder. Guards, which also make use of this behavior, are
    often code smell for areas where types with stronger guarantees will
    be a better choice. It helps guide us toward simpler functions, ones
    that are test-friendly and easier to reason about.

noInlineConfig:
    Some style guides suggest overriding rules when you have issues with
    them. The ability to arbitrarily override rules runs counter to what
    style enforcers are meant for. If there's issues with code style, it
    is better to discuss than decide individually how it should look.

* While TypeScript is better than using nothing, it is worth noting that
the type system is not sound, and that's by design. It's helpful, but it
means you can't fully trust your compiler to catch everything; you still
have to be vigilant! If you want a robust type system, my recommendation
would be an ML-family language, like Elm or PureScript.