Add unit testing
No unit tests? Surely you Jest! Well we do now with our initial test put
in place! We opted for Jest because we know, like Node, that it's a tool
used internally at Snap Mobile. Keen observers may also notice that when
we made evaluate() return an array for our test, we also said it will be
read-only. These kinds of considerations are important to make sure that
values aren't modified unknowingly when referenced in several places. It
becomes especially important in asynchronous and multi-threaded contexts
where data races can wreak havoc. No spooky action at a distance! On top
of that, values that don't change are easier to reason about and debug.