Initial parser
Now we start diving into the heart of this program: the parser! It's not
anything special right now; only numbers can be parsed. But we introduce
the concept of a parsing failure now. We also introduce fuzz testing for
the parser, making attempts to break it with invalid values or detect if
a parse isn't working like it should. In evaluate(), you might see we're
using .map() and think to yourself it would be inefficient if some value
was null and we still kept parsing. And you'd be right! But getting lost
in what-ifs like premature optimizations is a hinderance to getting work
done. When it becomes a performance problem or you have spare time, then
you address it.