Add operation evaluator
Now that we have the parser in place, we can implement an evaluator that
performs the operations. The evaluator is split into small parts to make
it easy to test. We use the individual evaluators in testing to ensure a
set of instructions produces the proper result on a stack, then use that
to also confirm that the parser is performing the right operations, too.
Since we parsed the input into a disjoint union that covers all possible
operations, we don't have to make sure operations coming in are valid or
not; we can focus solely on pure evaluation. The evaluator is now in the
internals since the public API now has an update function that calls the
parser and evaluator instead. Separation of code helps delineate what is
meant to be publicly consumed and what is exposed solely for testing. If
you wanted, you could put rules in place to disallow importing internals
from the outside.