Fix README for GitHub to use linked SVG instead
Having the SVG inlined would have meant that folks using screen readers
and other accessibility technologies could have "seen" the diagram with
the rest of the documentation, but GitHub does not allow inlining SVGs.
Understandable to an extent, but it limits our ability to make readable
documents for all kinds of people. Leaving a note that the original can
be read by accessibility technologies is the best we can do here.