Refactor individual evaluators to use Immutable.js
While migrating our code to persistent data structures, we want to focus
on containing changes to the smallest areas possible. When the behaviors
are finished and pass the original tests, we can then progressively move
these changes outward and update the tests accordingly. We also save the
original implementations as reference functions to use for testing while
we perform this refactor. Fuzz testing is good for cases like this!