Refactor evaluator to internally use Stack
We slowly expand our use of Stack out more to include evaluate() itself.
At its boundaries, it converts between array and Stack so the program is
unaware of any changes. This requires some changes to the original tests
to accommodate the transition. These changes are kept to a minimum, with
additional tests added as needed to make sure tests remain equivalent to
what they were before.