Refactor public API to use Stacks
The big moment: a breaking change to the public API of the library! It's
just a small change in our program to make it work though: make it fetch
the first item from the Stack instead of checking an array's index 0. It
wasn't so bad, huh? And since both the updater and evaluators use Stacks
all the way down, our tests can just make the migration steps the actual
steps!