Refactor parser to make a Stack of operations
Now that the evaluator has been taken care of, we'll refactor the parser
output to also use Stacks. Stacks, or rather linked lists, are useful in
creating a list of items where you're only really doing sequential reads
and operating on the head a lot.

I also ended up writing my own fuzz testing library for Jest because the
one I was using was a nightmare. No types and very confusing to use. Now
we have type-checked, composable continuation applicative functors. Yay!
Expect the existing tests to be rewritten in a future update. :)