Use jest-fuzzer instead of jest-fuzz
While fuzz testing is a great tool, jest-fuzz has a confusing API and no
types to help you or the compiler out. During the migration to immutable
types, I had made my own fuzzer with easily composable parts that is way
nicer to use and type checks really well. So we're rewriting all the old
jest-fuzz tests with my new Jest fuzzer and telling it to take a hike!