Truncate decimals longer than 2 places
With IEEE-754 floats, the techniques used to represent and do arithmetic
on them are fast, but they come at the cost of certain fraction patterns
having precision issues. We can fix the issues this causes when printing
to the console by processing the result into two parts, then joining the
parts together as a string. The issue with this implementation is that a
decimal is being truncated, not rounded.

This commit is the result of our interview's live coding session. Errors
are kept intact for demonstration purposes.