Add monadic binding operation to the Fuzzer
With the Fuzzer's `.map()` function, you can transform one type of value
into another directly. But what if you need to generate new, random data
inside a map? You previously had to make a direct call to `.generate()`,
which is a hack rather than a solution. So we introduced a way to take a
value and create a new Fuzzer from that. People familiar with functional
programming may recognize the pattern: our good friend, monadic binding!
It's called "bind", "chain", or even just `>>=`, but we use the Elm name
for it, "andThen".