Make the Fuzzer generator harder to use publicly
With the introduction of monadic binding, there's no need for the Fuzzer
`.generate()` method to be accessible publicly. The method uses a unique
symbol as its name now, one that isn't exported, to make accessing it in
outside code harder. The generator isn't private solely because it needs
to be usable by the fuzz testing functions in the library.